+++ classes/event/course_module_viewed.php
@@ -27,7 +27,7 @@ class course_module_viewed extends \core\event\course_module_viewed {
     /**
      * Initialise the event data.
      *
-     * @return void
+     * @return void No return value.
      */
     protected function init() {
         $this->data['crud'] = 'r';